server/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via
    #   httpx
    #   openai
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via tqdm
distro==1.9.0
    # via openai
django==5.0.1
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.3.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
djangorestframework-oauth==1.1.0
    # via -r requirements.in
exceptiongroup==1.2.0
    # via anyio
google-search-results==2.4.2
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.26.0
    # via openai
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
openai==1.11.1
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pydantic==2.6.1
    # via openai
pydantic-core==2.16.2
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2023.3.post1
    # via djangorestframework
requests==2.31.0
    # via google-search-results
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
sqlparse==0.4.4
    # via django
tqdm==4.66.1
    # via openai
typing-extensions==4.9.0
    # via
    #   anyio
    #   asgiref
    #   openai
    #   pydantic
    #   pydantic-core
tzdata==2023.4
    # via django
urllib3==2.2.0
    # via requests
